OTP-1944: Migration script to fill nReads and sequenceLength from FastQC results
Review #1952

+++ grails-app/services/de/dkfz/tbi/otp/ngsqc/FastqcUploadService.groovy
@@ -24,7 +24,7 @@ class FastqcUploadService {
         try {
             Map parsedFastqcFile = parseFastQCFile(fastqc, PROPERTIES_REGEX_TO_BE_PARSED)
             fastqc.dataFile.nReads = parsedFastqcFile["nReads"] as long
-            fastqc.dataFile.sequenceLength = parsedFastqcFile["sequenceLength"] as long
+            fastqc.dataFile.sequenceLength = parsedFastqcFile["sequenceLength"]
             fastqc.dataFile.save(flush: true, failOnError: true)
         } catch (final Throwable t) {
             throw new RuntimeException("Failed to load data from ${DATA_FILE_NAME} of ${fastqc} into the database.", t)
